move collectionview creation to view model Rename class to OfferDetailsViewModel
